renamed files and newhardware variable -> new_hardware
