+++ main.py
@@ -40,7 +40,7 @@ def run_simulator():
 )
 @click.option(
     "-n",
-    "--newhardware",
+    "--new_hardware",
     is_flag=True,
     default=False,
     help="Run on new hardware.",
@@ -54,9 +54,9 @@ def run_simulator():
     help="Run in test mode. This shortens the demo time and user input time "
     "for testing purposes.",
 )
-def run_kiosk(simulate, newhardware, testing):
+def run_kiosk(simulate, new_hardware, testing):
     """CLI command to run kiosk."""
-    kiosk.run(simulate, testing=testing, newhardware=newhardware)
+    kiosk.run(simulate, testing=testing, new_hardware=new_hardware)
 
 
 @cli.command("demo")
@@ -75,7 +75,7 @@ def run_kiosk(simulate, newhardware, testing):
 )
 @click.option(
     "-n",
-    "--newhardware",
+    "--new_hardware",
     is_flag=True,
     default=False,
     help="Run on new hardware.",
@@ -89,9 +89,9 @@ def run_kiosk(simulate, newhardware, testing):
     help="Run in test mode. This provides feedback for if your demo is "
     "running fast enough relative to the set frame rate.",
 )
-def run_demo(name, simulate, newhardware, testing):
+def run_demo(name, simulate, new_hardware, testing):
     """CLI command to run demo."""
-    demo.run(name, simulate, newhardware, testing=testing)
+    demo.run(name, simulate, new_hardware, testing=testing)
 
 
 @cli.command("test")
